Adjust configuration in preparation for unit-test
Add extra parameter to config.configure_app() to support passing in
config name for testing purpose
