+++ urlinfo/endpoint/resource_urlinfo.py
@@ -1,6 +1,7 @@
 from flask import Flask, Blueprint, current_app, jsonify, request
 import urllib
-from  ..model.urlinfo import UrlInfoParams, UrlInfo
+from ..database import UrlInfoRepository
+from ..model import UrlInfoParams, UrlInfo, db
 
 bp = Blueprint('resource_urlinfo', __name__)
 
@@ -11,6 +12,7 @@ def get(host_and_port, original_path_query_string=None):
     current_app.logger.info("urlinfoparam: " + str(url_info_param))
     url_info = url_info_param.to_urlinfo()
     current_app.logger.info("urlinfo: " + str(url_info))
-    found = UrlInfo.query.filter_by(url=url_info.url).first_or_404()
+    repository = UrlInfoRepository(db.session)
+    found = repository.get(url_info)
     current_app.logger.info("found: " + str(found))
     return jsonify({'malware': found.malware})
